+++ term1/seminar11_dynarray/classroom_tasks/code/5dynarray_header/dynarray.h
@@ -6,24 +6,27 @@
 
 typedef int Data;
 
-struct dynarray {
-	int capacity;
-	int size;
+struct dynarray 
+{
+	size_t capacity;
+	size_t size;
 	Data* values;
 };
 typedef struct dynarray Dynarray;
 
 
-void dynarray_init(Dynarray* d, int initial_capacity) 
+void dynarray_init(Dynarray* d, size_t initial_capacity) 
 {
 	d->capacity = initial_capacity;
 	d->size = 0;
 	d->values = (Data*)malloc(d->capacity * sizeof(Data));
 }
 
+
 void dynarray_push_back(Dynarray* d, Data x) 
 {
-	if (d->size >= d->capacity) {
+	if (d->size >= d->capacity) 
+    {
 		printf("Error! Dynarray is full! Can't push back any elements.\n");
 		exit(1);
 	}
@@ -31,28 +34,34 @@ void dynarray_push_back(Dynarray* d, Data x)
 	d->size += 1;
 }
 
-Data dynarray_get(const Dynarray* d, int index) 
+
+Data dynarray_get(const Dynarray* d, size_t index) 
 {
-	if (index < 0 || index >= d->size) {
+	if (index < 0 || index >= d->size) 
+    {
 		printf("Error! Index is out of range.\n");
 		exit(1);
 	}
 	return d->values[index];
 }
 
-void dynarray_set(Dynarray* d, int index, Data value) 
+
+void dynarray_set(Dynarray* d, size_t index, Data value) 
 {
-	if (index < 0 || index >= d->size) {
+	if (index < 0 || index >= d->size) 
+    {
 		printf("Error! Index is out of range.\n");
 		exit(1);
 	}
 	d->values[index] = value;
 }
 
+
 int dynarray_is_empty(const Dynarray* d) 
 {
 	return d->size == 0;
 }
+
 
 void dynarray_destroy(Dynarray* d) 
 {
